fix: 修改单测文件，添加vitest配置文件并改用vitest进行单元测试，并删除backup文件
