--- packages/shared/test/final-validation.mjs
@@ -1,222 +1,0 @@
-import * as fs from 'node:fs'
-import * as path from 'node:path'
-import { fileURLToPath } from 'node:url'
-// 最终验收测试：验证完整的 $r() 资源引用跳转功能
-import { parseResourceReference, ResourceResolver } from '../out/index.mjs'
-
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
-
-// 获取项目根目录
-const projectRoot = path.resolve(__dirname, '../../..')
-
-console.log('🚀 开始最终验收测试...')
-console.log('项目根目录:', projectRoot)
-
-// 测试用例：模拟实际的 $r() 使用场景
-const testCases = [
-  {
-    code: 'Text($r("app.string.app_name"))',
-    expected: {
-      type: 'string',
-      name: 'app_name',
-      file: 'string.json',
-      value: 'ArkTS Sample App',
-    },
-  },
-  {
-    code: '.fontColor($r(\'app.color.primary_color\'))',
-    expected: {
-      type: 'color',
-      name: 'primary_color',
-      file: 'color.json',
-      value: '#1976D2',
-    },
-  },
-  {
-    code: '.backgroundColor($r(`app.color.start_window_background`))',
-    expected: {
-      type: 'color',
-      name: 'start_window_background',
-      file: 'color.json',
-      value: '#FFFFFF',
-    },
-  },
-  {
-    code: 'Image($r("app.media.icon"))',
-    expected: {
-      type: 'media',
-      name: 'icon',
-      file: 'icon.png',
-      value: 'icon.png',
-    },
-  },
-  {
-    code: '.fontSize($r("app.float.page_text_font_size"))',
-    expected: {
-      type: 'float',
-      name: 'page_text_font_size',
-      file: 'float.json',
-      value: '50fp',
-    },
-  },
-]
-
-// 提取 $r() 调用的正则表达式
-function extractResourceReferences(code) {
-  const regex = /\$r\s*\(\s*['"`]([^'"`]+)['"`]\s*\)/g
-  const matches = []
-  let match
-
-  while ((match = regex.exec(code)) !== null) {
-    matches.push({
-      fullMatch: match[0],
-      resourceRef: match[1],
-      start: match.index,
-      end: match.index + match[0].length,
-    })
-  }
-
-  return matches
-}
-
-async function runValidationTests() {
-  console.log('\n=== 1. 初始化资源解析器 ===')
-
-  const resolver = new ResourceResolver(projectRoot)
-  await resolver.buildIndex()
-
-  const allResources = resolver.getAllResources()
-  console.log(`✅ 资源解析器初始化成功，找到 ${allResources.length} 个资源`)
-
-  console.log('\n=== 2. 测试 $r() 语法解析 ===')
-
-  let passedTests = 0
-  const totalTests = testCases.length
-
-  for (let i = 0; i < testCases.length; i++) {
-    const testCase = testCases[i]
-    console.log(`\n测试 ${i + 1}/${totalTests}: ${testCase.code}`)
-
-    try {
-      // 1. 提取 $r() 调用
-      const resourceRefs = extractResourceReferences(testCase.code)
-      if (resourceRefs.length === 0) {
-        console.log('❌ 无法提取 $r() 调用')
-        continue
-      }
-
-      const resourceRef = resourceRefs[0]
-      console.log(`   提取的资源引用: ${resourceRef.resourceRef}`)
-
-      // 2. 解析资源引用
-      const parsedRef = parseResourceReference(resourceRef.resourceRef)
-      if (!parsedRef) {
-        console.log('❌ 资源引用解析失败')
-        continue
-      }
-
-      console.log(`   解析结果: ${parsedRef.scope}.${parsedRef.type}.${parsedRef.name}`)
-
-      // 3. 查找资源位置
-      const resourceLocation = await resolver.resolveResourceReference(resourceRef.resourceRef)
-      if (!resourceLocation) {
-        console.log('❌ 无法找到资源位置')
-        continue
-      }
-
-      console.log(`   找到资源: ${path.basename(resourceLocation.uri)}`)
-      console.log(`   资源值: ${resourceLocation.value}`)
-
-      // 4. 验证结果
-      const fileName = path.basename(resourceLocation.uri)
-      const expectedFile = testCase.expected.file
-      const expectedValue = testCase.expected.value
-
-      if (fileName.includes(expectedFile) && resourceLocation.value === expectedValue) {
-        console.log('✅ 测试通过')
-        passedTests++
-      }
-      else {
-        console.log(`❌ 测试失败 - 预期文件: ${expectedFile}, 实际: ${fileName}`)
-        console.log(`   预期值: ${expectedValue}, 实际值: ${resourceLocation.value}`)
-      }
-    }
-    catch (error) {
-      console.log(`❌ 测试异常: ${error.message}`)
-    }
-  }
-
-  console.log(`\n=== 3. 测试结果汇总 ===`)
-  console.log(`通过: ${passedTests}/${totalTests}`)
-  console.log(`成功率: ${(passedTests / totalTests * 100).toFixed(1)}%`)
-
-  if (passedTests === totalTests) {
-    console.log('🎉 所有测试通过！$r() 资源引用跳转功能正常工作！')
-  }
-  else {
-    console.log('⚠️ 部分测试失败，需要进一步调试')
-  }
-}
-
-async function validateProjectStructure() {
-  console.log('\n=== 4. 验证项目结构完整性 ===')
-
-  // 检查关键文件是否存在
-  const keyFiles = [
-    'packages/shared/src/resource-resolver.ts',
-    'packages/language-server/src/services/resource-detector.service.ts',
-    'packages/language-server/src/services/resource-definition-simple.service.ts',
-    'sample/entry/src/main/resources/base/element/color.json',
-    'sample/entry/src/main/resources/base/element/string.json',
-    'sample/entry/src/main/resources/base/media/icon.png',
-    'sample/entry/src/main/ets/pages/Index.ets',
-  ]
-
-  let existingFiles = 0
-
-  for (const file of keyFiles) {
-    const filePath = path.join(projectRoot, file)
-    const exists = fs.existsSync(filePath)
-    console.log(`${exists ? '✅' : '❌'} ${file}`)
-    if (exists)
-      existingFiles++
-  }
-
-  console.log(`\n项目结构完整性: ${existingFiles}/${keyFiles.length} 文件存在`)
-
-  // 检查示例代码中的 $r() 使用
-  const indexFilePath = path.join(projectRoot, 'sample/entry/src/main/ets/pages/Index.ets')
-  if (fs.existsSync(indexFilePath)) {
-    const content = fs.readFileSync(indexFilePath, 'utf-8')
-    const resourceRefs = extractResourceReferences(content)
-    console.log(`\n示例代码中发现 ${resourceRefs.length} 个 $r() 调用:`)
-    resourceRefs.forEach((ref, index) => {
-      console.log(`  ${index + 1}. ${ref.resourceRef}`)
-    })
-  }
-}
-
-async function runFinalValidation() {
-  try {
-    await runValidationTests()
-    await validateProjectStructure()
-
-    console.log('\n🏆 最终验收测试完成！')
-    console.log('\n📋 功能确认清单:')
-    console.log('✅ 资源引用解析 - parseResourceReference()')
-    console.log('✅ 资源索引构建 - ResourceResolver.buildIndex()')
-    console.log('✅ 资源位置查找 - ResourceResolver.resolveResourceReference()')
-    console.log('✅ 多种资源类型支持 - color, string, media, float')
-    console.log('✅ 多种引号格式支持 - \'\', "", ``')
-    console.log('✅ 多模块项目支持 - entry, sampleLibrary')
-    console.log('✅ 错误处理和边界条件')
-    console.log('✅ 项目结构完整性')
-  }
-  catch (error) {
-    console.error('❌ 最终验收测试失败:', error)
-  }
-}
-
-// 运行最终验收测试
-runFinalValidation()
